+++ modules/eks/ray-cluster/ray_cluster_stack.py
@@ -6,6 +6,7 @@ from typing import Any, Dict, List, Optional, cast
 
 from aws_cdk import Stack, Tags
 from aws_cdk import aws_eks as eks
+from aws_cdk import aws_iam as iam
 from aws_cdk.lambda_layer_kubectl_v29 import KubectlV29Layer
 from constructs import Construct, IConstruct
 
@@ -23,6 +24,7 @@ class RayCluster(Stack):
         module_name: str,
         eks_cluster_name: str,
         eks_admin_role_arn: str,
+        eks_handler_role_arn: str,
         eks_openid_connect_provider_arn: str,
         namespace_name: str,
         service_account_name: str,
@@ -62,6 +64,7 @@ class RayCluster(Stack):
         provider = eks.OpenIdConnectProvider.from_open_id_connect_provider_arn(
             self, "OIDCProvider", eks_openid_connect_provider_arn
         )
+        handler_role = iam.Role.from_role_arn(self, "HandlerRole", eks_handler_role_arn)
 
         eks_cluster = eks.Cluster.from_cluster_attributes(
             self,
@@ -70,6 +73,7 @@ class RayCluster(Stack):
             open_id_connect_provider=provider,
             kubectl_role_arn=eks_admin_role_arn,
             kubectl_layer=KubectlV29Layer(self, "Kubectlv29Layer"),
+            kubectl_lambda_role=handler_role,
         )
 
         volumes = [