+++ modules/eks/ray-cluster/tests/test_stack.py
@@ -33,6 +33,7 @@ def ray_cluster_stack(stack_defaults) -> cdk.Stack:
 
     eks_cluster_name = "cluster"
     eks_cluster_admin_role_arn = "arn:aws:iam::123456789012:role/eks-testing-XXXXXX"
+    eks_handler_role_arn = "arn:aws:iam::123456789012:role/eks-testing-XXXXXX"
     eks_oidc_arn = "arn:aws:iam::123456789012:oidc-provider/oidc.eks.us-east-1.amazonaws.com/id/XXXXXXXX"
     namespace = "namespace"
     service_account_name = "acc"
@@ -61,6 +62,7 @@ def ray_cluster_stack(stack_defaults) -> cdk.Stack:
         module_name=mod_name,
         eks_cluster_name=eks_cluster_name,
         eks_admin_role_arn=eks_cluster_admin_role_arn,
+        eks_handler_role_arn=eks_handler_role_arn,
         eks_openid_connect_provider_arn=eks_oidc_arn,
         namespace_name=namespace,
         service_account_name=service_account_name,